GIX-1017: CR changes
