+++ frontend/src/lib/constants/routes.constants.ts
@@ -1,5 +1,3 @@
-import { ENABLE_SNS } from "./environment.constants";
-
 export enum AppPath {
   Authentication = "/",
   LegacyAccounts = "/#/accounts",
@@ -17,24 +15,5 @@ export enum AppPath {
   ProjectDetail = "/#/u",
   NeuronDetail = "/#/u/:rootCanisterId/neuron",
 }
-
-export const paths = {
-  neuronDetail: (rootCanisterId: string) =>
-    ENABLE_SNS
-      ? `${CONTEXT_PATH}/${rootCanisterId}/neuron`
-      : AppPath.LegacyNeuronDetail,
-  neurons: (rootCanisterId: string) =>
-    ENABLE_SNS
-      ? `${CONTEXT_PATH}/${rootCanisterId}/neurons`
-      : AppPath.LegacyNeurons,
-  accounts: (rootCanisterId: string) =>
-    ENABLE_SNS
-      ? `${CONTEXT_PATH}/${rootCanisterId}/accounts`
-      : AppPath.LegacyAccounts,
-  wallet: (rootCanisterId: string) =>
-    ENABLE_SNS
-      ? `${CONTEXT_PATH}/${rootCanisterId}/wallet`
-      : AppPath.LegacyWallet,
-};
 
 export const CONTEXT_PATH = "/#/u";
